% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{absolute.order}
\alias{absolute.order}
\title{absolute magnitude order}
\usage{
absolute.order(beta)
}
\arguments{
\item{beta}{a vector of regression coefficients}
}
\description{
This function extracts the absolute value order from any fit.
}
\examples{
### generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

### glmnet fit
library(glmnet)
beta.lasso = coef(cv.glmnet(X, y))[-1]
lasso.order = mr.ash.alpha:::absolute.order(beta.lasso)

### ncvreg fit
library(ncvreg)
beta.scad = c(coef(cv.ncvreg(X, y))[-1])
scad.order = mr.ash.alpha:::absolute.order(beta.scad)

}
