% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{coef.mr.ash}
\alias{coef.mr.ash}
\title{extract regression coefficients from mr_ash fit}
\usage{
\method{coef}{mr.ash}(object, ...)
}
\arguments{
\item{object}{a mr_ash fit}
}
\value{
a p+1 vector, the first element being an intercept, and the
  remaining p elements being estimated regression coefficients
  
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X %*% beta + rnorm(n)

## fit mr.ash
fit.mr.ash  = mr.ash(X, y)

## coefficient
coef.mr.ash = coef(fit.mr.ash)
intercept   = coef.mr.ash[1]
beta        = coef.mr.ash[-1]
}
\description{
extract regression coefficients from mr_ash fit
}
