% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{coef.mr_ash}
\alias{coef.mr_ash}
\title{extract regression coefficients from mr_ash fit}
\usage{
\method{coef}{mr_ash}(object, ...)
}
\arguments{
\item{object}{a mr_ash fit}
}
\value{
a p+1 vector, the first element being an intercept, and the remaining p elements being estimated regression coefficients
}
\description{
extract regression coefficients from mr_ash fit
}
