% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{computelfsrmix}
\alias{computelfsrmix}
\title{compute local false discovery rate}
\usage{
computelfsrmix(alpha, mu, s)
}
\description{
----------------------------------------------------------------------
Compute the local false sign rate (LFSR) for each variable. This
assumes that the first mixture component is a "spike" (that is, a
normal density with a variance approaching zero).
----------------------------------------------------------------------
}
