% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.full.posterior}
\alias{get.full.posterior}
\title{get q}
\usage{
get.full.posterior(fit)
}
\arguments{
\item{fit}{An Mr.ASH fit obtained by running \code{mr.ash}. See \sQuote{Examples}.}
}
\value{
A list object with the following elements:

\item{phi}{A posterior component probability matrix of dimension (p,K);
each row is a vector posterior mixture proportions corresponding to
each regression coefficient.}

\item{m}{A posterior component mean matrix of dimension (p,K);
each row is a vector of posterior component means corresponding to
each regression coefficient.}

\item{s2}{A posterior component variance matrix of dimension (p,K);
each row is a vector of posterior component variances corresponding to
each regression coefficient.}
}
\description{
The \code{get.full.posterior} function recovers the estimated
variational posterior, which maximizes the ELBO F(q,g,sigma2). The \code{mr.ash}
function does not store the full posterior for a faster implementation, thus
this function \code{get.full.posterior} recovers the full posterior from the
Mr.ASH fit.
}
\examples{
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

## fit mr.ash
fit.mr.ash  = mr.ash(X, y)

## recover full posterior
full.post   = get.full.posterior(fit.mr.ash)

}
