% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logspace}
\alias{logspace}
\title{----------------------------------------------------------------------}
\usage{
logspace(a = 0, b, length = 10)
}
\description{
----------------------------------------------------------------------
}
