% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ash.R
\name{mr_ash}
\alias{mr_ash}
\title{Something}
\usage{
mr_ash(X, y, Z = NULL, sa2 = NULL, K = 10, method = "caisa_em",
  max.iter = 1000, min.iter = 1, stepsize = 1, tol = list(),
  outputlevel = 1, update.sigma = TRUE, sigma2 = NULL,
  beta.init = NULL, standardize = FALSE, intercept = TRUE,
  verbose = TRUE)
}
\arguments{
\item{X}{Describe X here. Sparse matrix? Dense?}

\item{y}{Describe y here.}
}
\value{
Describe return value here.
}
\description{

}
\details{
Add details here.
}
\examples{

set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y = X \%*\% beta + rnorm(n)

fit.mr_ash_em  = mr_ash(X,y, method = "caisa_em")

fit.mr_ash_g   = mr_ash(X,y, method = "caisa_g")

fit.mr_ash_acc = mr_ash(X,y, method = "caisa_acc")

Xnew        = matrix(rnorm(n*p),n,p)
ynew        = predict.mr_ash(fit.mr_ash, Xnew)

head(data.frame("caisa_em" = ynew_em, "caisa_g" = ynew_g, "caisa_acc" = ynew_acc, "true" = ynew))

}
