% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ash_order.R
\name{mr_ash_order}
\alias{mr_ash_order}
\title{Something}
\usage{
mr_ash_order(X, y, Z = NULL, sa2 = NULL, order.method = "1:p",
  o = NULL, K = 10, max.iter = 1000, min.iter = 1, stepsize = 1,
  tol = list(), outputlevel = 1, update.sigma = TRUE,
  sigma2 = NULL, beta.init = NULL, standardize = FALSE,
  intercept = TRUE, verbose = TRUE)
}
\arguments{
\item{X}{Describe X here. Sparse matrix? Dense?}

\item{y}{Describe y here.}
}
\value{
Describe return value here.
}
\description{

}
\details{
Add details here.
}
\examples{

set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y = X \%*\% beta + rnorm(n)

fit.mr_ash  = mr_ash_order(X,y, order.method = "random")

# DO NOT RUN

# fit.mr_ash  = mr_ash_order(X,y, order.method = "normal")

# fit.mr_ash  = mr_ash_order(X,y, order.method = "given", o = rep(0:(p-1), max.iter))

Xnew        = matrix(rnorm(n*p),n,p)
ynew        = predict.mr_ash(fit.mr_ash, Xnew)


}
