% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{path.order}
\alias{path.order}
\title{regularization path order}
\usage{
path.order(fit)
}
\arguments{
\item{beta}{a glmnet fit, or a ncvreg fit}
}
\description{
This function extracts the path order from penalized
regression fit.
}
\examples{
### generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

### glmnet fit
library(glmnet)
fit.lasso = glmnet(X, y)
lasso.order = path.order(fit.lasso)

### ncvreg fit
library(ncvreg)
fit.scad = ncvreg(X, y)
scad.order = path.order(fit.scad)

}
