% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{post_median}
\alias{post_median}
\title{----------------------------------------------------------------------
Some miscellaneuous auxiliary functions are listed above.
Some functions are directly borrowed from the R package varbvs
https://github.com/pcarbo/varbvs
----------------------------------------------------------------------}
\usage{
post_median(beta, r, d, Phi, sa2, thresh = 0.9)
}
\description{
----------------------------------------------------------------------
Some miscellaneuous auxiliary functions are listed above.
Some functions are directly borrowed from the R package varbvs
https://github.com/pcarbo/varbvs
----------------------------------------------------------------------
}
