% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{predict.caisa}
\alias{predict.caisa}
\title{----------------------------------------------------------------------}
\usage{
\method{predict}{caisa}(fit, X)
}
\description{
----------------------------------------------------------------------
}
