% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{proj_l0}
\alias{proj_l0}
\title{----------------------------------------------------------------------}
\usage{
proj_l0(b, K = 10)
}
\description{
----------------------------------------------------------------------
}
