% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_covariate}
\alias{remove_covariate}
\title{remove covariate effects}
\usage{
remove_covariate(X, y, Z, standardize = FALSE, intercept = TRUE)
}
\description{
This function regresses the effect of Z out from X and y.
In other words, X and y will be projected into the space orthogonal to Z.
}
