% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{univar.order}
\alias{univar.order}
\title{univariate order}
\usage{
univar.order(X, y)
}
\arguments{
\item{X}{An input design matrix.}

\item{y}{A vector of response variables.}
}
\description{
This function extracts the univariate regression coefficient order
}
\examples{
### generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

univ.order = univar.order(X,y)

}
