% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmixopt.R
\name{varmixopt}
\alias{varmixopt}
\title{Something}
\usage{
varmixopt(X, y, Z = NULL, sa2 = NULL, K = 10, method = "caisa",
  max.iter = 1000, min.iter = 1, stepsize = 1, tol = list(),
  outputlevel = 1, update.sigma = TRUE, sigma2 = NULL,
  beta.init = NULL, standardize = FALSE, intercept = TRUE,
  verbose = TRUE)
}
\arguments{
\item{X}{Describe X here. Sparse matrix? Dense?}

\item{y}{Describe y here.}
}
\value{
Describe return value here.
}
\description{

}
\details{
Add details here.
}
