% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.ash.R
\name{mr.ash}
\alias{mr.ash}
\title{Mr.ASH (Multiple Regression with Adaptive Shrikage)}
\usage{
mr.ash(X, y, Z = NULL, sa2 = c((2^((0:19)/20) - 1)^2, NULL), K = 20,
  method = c("caisa", "accelerate", "block", "sigma_scaled",
  "sigma_indep"), max.iter = 1000, min.iter = 1, beta.init = NULL,
  update.pi = TRUE, pi = NULL, update.sigma = TRUE, sigma2 = NULL,
  update.order = NULL, standardize = FALSE, intercept = TRUE,
  mixsqpiter = 5, mode = FALSE, tol = set_default_tolerance(),
  verbose = TRUE)
}
\arguments{
\item{X}{Describe X here. Sparse matrix? Dense?}

\item{y}{Describe y here.}
}
\value{
Describe return value here.
}
\description{

}
\details{
Add details here.
}
\examples{

set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

fit.mr.ash  = mr.ash(X,y, method = "caisa")
# fit.mr.ash  = mr.ash(X,y, method = "update_g")
# fit.mr.ash  = mr.ash(X,y, method = "accelerate")

Xnew        = matrix(rnorm(n*p),n,p)
ynew        = Xnew \%*\% beta + rnorm(n)
ypred       = predict(fit.mr.ash, Xnew)
rmse        = norm(ynew - ypred, '2') / sqrt(n)
betahat     = fit.mr.ash$beta

}
