% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{predict.mr.ash}
\alias{predict.mr.ash}
\title{predict future observations or extract coefficients from mr_ash fit}
\usage{
\method{predict}{mr.ash}(object, newx = NULL, type = c("response", "coefficients"), ...)
}
\arguments{
\item{object}{a mr_ash fit}

\item{newx}{a new value for X at which to do predictions}

\item{type}{if this is coefficients, then calls coef.susie}
}
\description{
predict future observations or extract coefficients from mr_ash fit
}
\examples{
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

## fit mr.ash
fit.mr.ash  = mr.ash(X, y)

## predict
Xnew        = matrix(rnorm(n*p),n,p)
ypred       = predict(fit.mr.ash, Xnew)


}
