% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_covariate}
\alias{remove_covariate}
\title{remove covariate effects}
\usage{
remove_covariate(X, y, Z, standardize = FALSE, intercept = TRUE)
}
\description{
----------------------------------------------------------------------
Some miscellaneuous auxiliary functions are listed above.
Some functions are directly borrowed from the R package varbvs
https://github.com/pcarbo/varbvs
----------------------------------------------------------------------
----------------------------------------------------------------------
y = Z * alpha + X * beta + epsilon
remove Z and project X and y onto space orthogonal to Z
add intercept (a vecror of ones) to Z if intercept == TRUE
standardize X if standardize == TRUE
----------------------------------------------------------------------
}
